deleted unused classes (usaged of ArrowIncrementDecrementOverlay), added Animator to Arrow..Overlay to allow Randomize-Items to play animation when the action is executed, closes #3503
