--- projects/epc/playground/src/proxies/hwui/descriptive-layouts/concrete/menu/menu-items/ParameterItem.h
@@ -1,14 +1,0 @@
-#pragma once
-#include "EditorItem.h"
-
-class Parameter;
-
-class ParameterItem : public EditorItem
-{
- public:
-  ParameterItem(Parameter *parameter, const Rect &r);
-  MenuOverlay *createOverlay() override;
-
- private:
-  Parameter *const m_parameter;
-};
