--- projects/epc/playground/src/proxies/hwui/descriptive-layouts/concrete/menu/menu-overlays/ParameterOverlay.h
@@ -1,17 +1,0 @@
-#pragma once
-#include "ArrowIncrementDecrementOverlay.h"
-
-class Parameter;
-class ParameterValueLabel;
-
-class ParameterOverlay : public ArrowIncrementDecrementOverlay
-{
- public:
-  explicit ParameterOverlay(const Rect& rect, Parameter* parameter);
-  void onLeft(bool down) override;
-  void onRight(bool down) override;
-
- private:
-  ParameterValueLabel* m_label;
-  Parameter* const m_parameter;
-};
